Fold conversion from RemoteAuthentication -> pcap_rmauth in RemotePcap into RemoteAuthentication for ease of following the code.
Note that this is an API breaking change.
